postmarketos-initramfs: search for root partition in mount_subpartitions
Currently, mount_subpartitions checks for the pmOS_boot partition to
exist. However it might be the case that the pmOS_boot partition is on a
dedicated partition while the pmOS_root partition is in a subpartition.
This could happen when migrating to generic UEFI device support on an
existing installation.

To ensure that we definitely find the rootfs if available, adjust the
check.

Signed-off-by: Caleb Connolly <caleb@connolly.tech>